refactor: ComposedBenchmarks live outside tasks submodule
